refactor: move invoicing to cloudflare workflows
