--- apps/billing/src/lib/clickhouse.ts
@@ -1,4 +1,0 @@
-import { ClickHouse } from "@unkey/clickhouse";
-import { env } from "./env";
-
-export const clickhouse = new ClickHouse({ url: env().CLICKHOUSE_URL });
