+++ apps/workflows/wrangler.toml
@@ -26,5 +26,13 @@ binding = "COUNT_KEYS"
 # this is class that extends the Workflow class in src/index.ts
 class_name = "CountKeys"
 
+
+[[workflows]]
+# name of your workflow
+name = "invoicing"
+# binding name env.MYWORKFLOW
+binding = "INVOICING"
+# this is class that extends the Workflow class in src/index.ts
+class_name = "Invoicing"
 [triggers]
-crons = ["* * * * *", "0 0 * * *"]
+crons = ["* * * * *", "0 0 * * *", "0 13 1 * *"]
